Show symlinks in wallpaper list.
In the BunsenLabs wallpaper directory there is a symlink named "default" which points to the current default wallpaper, and is used in the initial BL settings. That symlink does not appear in the xwwall button list, which offers only the real files.

Whether symlinks should also appear in the wallpaper list is your final call, but they could easily be added with the --dereference option to file on line 381.